requirements.txt
click==6.6
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
gunicorn==20.0.4 
numpy==1.19.4
openpyxl==3.0.5
pandas==1.1.4
pandas-profiling==2.9.0
requests==2.24.0
scikit-learn==0.23.2
sklearn-pandas==2.0.3 
scipy==1.5.3
xlrd==1.2.0
Werkzeug==0.15.4